+++ sofa/engine.lua
@@ -4,13 +4,15 @@ local utils = require("sofa.utils")
 local engine = {}
 
 ---@class ExecutionEngine
+---@field private shell string
 local ExecutionEngine = {}
 
 ---create a new execution engine based on the configuration
----@param _config table
+---@param config table
 ---@return ExecutionEngine
-function ExecutionEngine:new(_config)
+function ExecutionEngine:new(config)
   local o = {}
+  o.shell = config.shell or "bash"
   setmetatable(o, self)
   self.__index = self
   return o
@@ -19,8 +21,7 @@ end
 ---run a command using the execution engine
 ---@param command string
 function ExecutionEngine:run(command)
-  -- TODO: make shell configurable
-  local cmd = string.format("bash -c '%s'", utils.escape_quotes(command))
+  local cmd = string.format("%s -c '%s'", self.shell, utils.escape_quotes(command))
   os.execute(cmd)
 end