+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
+Features:
+- Support setting the execution shell when executing command in non-interactive mode.
+
 Improvements:
 - Also provide non-interactive commands when launched with `-i`.
 - Show default in the command substitution if set.
+- Allow to cleanly exit by aborting rofi.
 
 Fixes:
 - Correctly implement exclusive choices selection.